# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para gestão de ordens de serviço, clientes, veículos, serviços e estoque de peças.",
        "title": "Oficina Mecânica API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "TechChalleng",
            "email": "suporte@oficina.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/ordens-servico": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "ordens-servico"
                ],
                "summary": "Listar ordens de serviço",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.OrdemServicoResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cria uma nova OS no status \"Recebida\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordens-servico"
                ],
                "summary": "Criar ordem de serviço",
                "parameters": [
                    {
                        "description": "Dados da OS",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.CriarOSRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.OrdemServicoResponse"
                        }
                    }
                }
            }
        },
        "/ordens-servico/metricas/tempo-medio": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna o tempo médio em horas entre criação e entrega",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordens-servico"
                ],
                "summary": "Tempo médio de execução das OS",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.TempoMedioResponse"
                        }
                    }
                }
            }
        },
        "/ordens-servico/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "ordens-servico"
                ],
                "summary": "Buscar ordem de serviço por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da OS",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.OrdemServicoResponse"
                        }
                    }
                }
            }
        },
        "/ordens-servico/{id}/avancar": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "ordens-servico"
                ],
                "summary": "Avançar status da OS automaticamente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da OS",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.OrdemServicoResponse"
                        }
                    }
                }
            }
        },
        "/ordens-servico/{id}/pecas": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Verifica saldo de estoque antes de incluir",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordens-servico"
                ],
                "summary": "Adicionar peça à OS",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da OS",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Peça",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.AdicionarPecaOSRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.OrdemServicoResponse"
                        }
                    }
                }
            }
        },
        "/ordens-servico/{id}/servicos": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordens-servico"
                ],
                "summary": "Adicionar serviço à OS",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da OS",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Serviço",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.AdicionarServicoOSRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.OrdemServicoResponse"
                        }
                    }
                }
            }
        },
        "/ordens-servico/{id}/status": {
            "get": {
                "description": "Rota pública para o cliente acompanhar o progresso da OS sem autenticação",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordens-servico"
                ],
                "summary": "Consulta pública de status da OS",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da OS",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.StatusOSResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Transição controlada por máquina de estados. Ao \"Finalizada\", baixa estoque automaticamente.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ordens-servico"
                ],
                "summary": "Atualizar status da OS",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da OS",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Novo status",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.AtualizarStatusOSRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.OrdemServicoResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_techchalleng_oficina_internal_application_dto.AdicionarPecaOSRequest": {
            "type": "object",
            "required": [
                "peca_id",
                "quantidade"
            ],
            "properties": {
                "peca_id": {
                    "type": "string"
                },
                "quantidade": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "github_com_techchalleng_oficina_internal_application_dto.AdicionarServicoOSRequest": {
            "type": "object",
            "required": [
                "servico_id"
            ],
            "properties": {
                "servico_id": {
                    "type": "string"
                }
            }
        },
        "github_com_techchalleng_oficina_internal_application_dto.AtualizarStatusOSRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
        "github_com_techchalleng_oficina_internal_application_dto.CriarOSRequest": {
            "type": "object",
            "required": [
                "cliente_id",
                "veiculo_id"
            ],
            "properties": {
                "cliente_id": {
                    "type": "string"
                },
                "observacoes": {
                    "type": "string"
                },
                "veiculo_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_techchalleng_oficina_internal_application_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "github_com_techchalleng_oficina_internal_application_dto.ItemPecaOSResponse": {
            "type": "object",
            "properties": {
                "nome": {
                    "type": "string"
                },
                "peca_id": {
                    "type": "string"
                },
                "quantidade": {
                    "type": "integer"
                },
                "valor_unitario": {
                    "type": "number"
                }
            }
        },
        "github_com_techchalleng_oficina_internal_application_dto.ItemServicoOSResponse": {
            "type": "object",
            "properties": {
                "nome": {
                    "type": "string"
                },
                "servico_id": {
                    "type": "string"
                },
                "valor_mao_obra": {
                    "type": "number"
                }
            }
        },
...
        "github_com_techchalleng_oficina_internal_application_dto.OrdemServicoResponse": {
            "type": "object",
            "properties": {
                "atualizado_em": {
                    "type": "string"
                },
                "cliente_id": {
                    "type": "string"
                },
                "criado_em": {
                    "type": "string"
                },
                "entregue_em": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "observacoes": {
                    "type": "string"
                },
                "pecas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.ItemPecaOSResponse"
                    }
                },
                "servicos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_techchalleng_oficina_internal_application_dto.ItemServicoOSResponse"
                    }
                },
                "status": {
                    "type": "string"
                },
                "valor_total": {
                    "type": "number"
                },
                "veiculo_id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_techchalleng_oficina_internal_application_dto.StatusOSResponse": {
            "type": "object",
            "properties": {
                "criado_em": {
                    "type": "string"
                },
                "entregue_em": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "valor_total": {
                    "type": "number"
                }
            }
        },
        "github_com_techchalleng_oficina_internal_application_dto.TempoMedioResponse": {
            "type": "object",
            "properties": {
                "tempo_medio_horas": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Insira o token JWT no formato: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}